+++ fake_user_agent/thread_version.py
@@ -1,6 +1,5 @@
 """Generate fake useragent using threading. This is only for offering another solution. Not actually called in the published pkg"""
 import os
-import glob
 import json
 import random
 from time import sleep
@@ -84,6 +83,7 @@ def load_by_threadpool(use_cache_server=True):
 
 
 def write(path, data):
+    rm_tempfile()
     with open(path, encoding="utf-8", mode="wt") as f:
         dumped = json.dumps(data)
         f.write(dumped)
@@ -96,9 +96,12 @@ def read(path):
 
 
 def rm_tempfile():
-    tempfile_list = glob.glob(os.path.join(settings.DB_DIR, "fake_useragent_*"))
-    for i in tempfile_list:
-        os.remove(i)
+    tempfile_list = settings.TEMP_FILE
+    if tempfile_list:
+        for i in tempfile_list:
+            os.remove(i)
+    else:
+        return
 
 
 def random_choose(browser, data):
@@ -123,8 +126,8 @@ def user_agent(browser=None, use_tempfile=True):
         if browser not in list(settings.BROWSERS.keys()):
             raise FakeUserAgentError("This browser is not supported.")
 
-    if os.path.isfile(settings.DB):
-        data = read(settings.DB)
+    if settings.TEMP_FILE:
+        data = read(settings.TEMP_FILE[-1])
         return random_choose(browser, data)
 
     else:
